backend/python-rgcn/requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
neo4j>=5.15.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0

